Make HANNA a package
