--- database/Activity/HANNA/phi_2_b.csv
@@ -1,1 +1,0 @@
-2.406278625130653381e-02
