--- database/Activity/HANNA/phi_2_w.csv
@@ -1,1 +1,0 @@
--2.588994204998016357e-01,1.110284477472305298e-01,-2.497833520174026489e-01,1.179642081260681152e-01,-2.413461953401565552e-01,1.406791806221008301e-01,7.329827547073364258e-02,-1.739050745964050293e-01,-1.469288108637556434e-04,-2.572799623012542725e-01,-2.184250801801681519e-01,-1.739730387926101685e-01,-1.247205138206481934e-01,1.234945803880691528e-01,-1.854458153247833252e-01,1.209850013256072998e-01,7.175085693597793579e-02,-2.099588811397552490e-01,6.591714918613433838e-02,-2.772944569587707520e-01,-2.127214670181274414e-01,-2.098390609025955200e-01,1.175448372960090637e-01,1.051775738596916199e-01,1.216908693313598633e-01,-2.600685060024261475e-01,-1.779023408889770508e-01,-2.909160554409027100e-01,1.468326747417449951e-01,-1.667334437370300293e-01,1.572261303663253784e-01,8.996020257472991943e-02,-2.156818211078643799e-01,4.508634656667709351e-02,1.317703127861022949e-01,8.760061115026473999e-02,-1.544055193662643433e-01,1.128358840942382812e-01,-2.109066098928451538e-01,-1.487377882003784180e-01,7.590584456920623779e-02,-1.713133156299591064e-01,-3.117692470550537109e-01,1.336891800165176392e-01,7.362193614244461060e-02,-2.386992573738098145e-01,1.347091048955917358e-01,-2.013245075941085815e-01,-2.272962331771850586e-01,1.222808584570884705e-01,1.818726211786270142e-01,1.015873923897743225e-01,-1.744805723428726196e-01,-2.083844989538192749e-01,-1.481742411851882935e-01,-2.737368941307067871e-01,8.546546101570129395e-02,1.107692196965217590e-01,-2.317997515201568604e-01,6.139200925827026367e-02,1.362721025943756104e-01,-2.316985875368118286e-01,5.011718720197677612e-02,-2.417234480381011963e-01,-2.308850288391113281e-01,1.644366681575775146e-01,-2.391201704740524292e-01,-1.953793317079544067e-01,-2.604509592056274414e-01,3.602461516857147217e-02,1.314551979303359985e-01,-1.190138980746269226e-01,5.440299585461616516e-02,1.190554574131965637e-01,6.670927163213491440e-03,1.019747480750083923e-01,-1.669223308563232422e-01,-1.911280453205108643e-01,1.759074479341506958e-01,1.294234395027160645e-01,-2.148957848548889160e-01,-4.439926147460937500e-01,-2.202425301074981689e-01,9.249615669250488281e-02,6.739849597215652466e-02,1.650258153676986694e-01,-2.362324297428131104e-01,1.091802194714546204e-01,-2.947779297828674316e-01,1.038336828351020813e-01,-2.439362853765487671e-01,-1.881664991378784180e-01,6.045620515942573547e-02,1.549287438392639160e-01,5.895221605896949768e-02,-2.222333699464797974e-01
