--- database/Activity/HANNA/theta_1_b.csv
@@ -1,96 +1,0 @@
--3.677943050861358643e-01
--1.948349475860595703e-01
--3.274458348751068115e-01
--2.574045062065124512e-01
--1.655054390430450439e-01
--3.641201853752136230e-01
--3.302614986896514893e-01
--5.470842719078063965e-01
--2.922910824418067932e-02
--5.281679034233093262e-01
--4.239531159400939941e-01
--4.728624224662780762e-01
--3.956952095031738281e-01
--3.254536092281341553e-01
--4.827946126461029053e-01
--4.433381259441375732e-01
--4.597900807857513428e-01
-1.124040633440017700e-01
--1.023616790771484375e-01
--7.909586280584335327e-02
--2.289993017911911011e-01
--3.586184978485107422e-01
--1.549386680126190186e-01
--5.822034478187561035e-01
--2.928400933742523193e-01
--3.516216874122619629e-01
--2.919531166553497314e-01
--6.511517167091369629e-01
--2.370001077651977539e-01
--1.613234877586364746e-01
--4.129924476146697998e-01
--3.622640967369079590e-01
--3.979320228099822998e-01
--4.141536653041839600e-01
--4.538420438766479492e-01
--5.563888549804687500e-01
--1.981357187032699585e-01
--4.129261672496795654e-01
--4.334511160850524902e-01
--2.827472388744354248e-01
--1.186183691024780273e-01
--3.743814826011657715e-01
--5.585865974426269531e-01
--4.180952012538909912e-01
--4.154843986034393311e-01
--2.389999181032180786e-01
--2.638523578643798828e-01
--4.264073669910430908e-01
--2.737129330635070801e-01
--2.226671725511550903e-01
--4.526865780353546143e-01
--2.926479876041412354e-01
--5.836908221244812012e-01
--6.171780228614807129e-01
--3.901229500770568848e-01
--1.567516028881072998e-01
--2.458352297544479370e-01
--2.589947283267974854e-01
--3.526111245155334473e-01
--3.113025128841400146e-01
--5.328215956687927246e-01
--3.253822922706604004e-01
--2.846125066280364990e-01
-1.905919909477233887e-01
-1.862462311983108521e-01
--5.350244641304016113e-01
--4.941167831420898438e-01
--2.223552316427230835e-01
--5.124427080154418945e-01
--2.368766963481903076e-01
--4.107604622840881348e-01
--5.221566557884216309e-01
--2.872848212718963623e-01
--3.616807460784912109e-01
--7.744724750518798828e-01
--3.195562958717346191e-01
--4.797289669513702393e-01
--2.955931425094604492e-01
--1.263094395399093628e-01
--4.355602562427520752e-01
--4.270291030406951904e-01
-4.804476723074913025e-02
--8.052706718444824219e-01
--2.674348950386047363e-01
--2.335024625062942505e-01
--3.606138825416564941e-01
--3.262173533439636230e-01
--2.510459423065185547e-01
--2.405110150575637817e-01
--4.745654463768005371e-01
--7.932834029197692871e-01
--4.244997203350067139e-01
--3.705306351184844971e-01
--4.078170359134674072e-01
--3.132100403308868408e-01
--3.594588935375213623e-01
