+++ docs/src/eos/activity.md
@@ -43,6 +43,7 @@ Clapeyron.tcPRWilsonRes
 Clapeyron.COSMOSAC02
 Clapeyron.COSMOSAC10
 Clapeyron.COSMOSACdsp
+Clapeyron.HANNA
 ```
