+++ lib/ClapeyronHANNA/data/HANNA/alpha_1_b.csv
@@ -1,0 +1,96 @@
+1.113308519124984741e-01
+-3.546121343970298767e-02
+-2.988076396286487579e-02
+1.303256005048751831e-01
+-4.425702989101409912e-02
+1.375684738159179688e-01
+-5.593518167734146118e-02
+-9.229625761508941650e-02
+8.451537787914276123e-02
+4.212076961994171143e-02
+-1.848700791597366333e-01
+-3.610199317336082458e-02
+8.203893899917602539e-03
+2.458231523633003235e-02
+3.563912212848663330e-02
+2.128550708293914795e-01
+-5.403746291995048523e-02
+-1.766296476125717163e-01
+2.704654075205326080e-02
+-2.685149386525154114e-02
+-2.528327405452728271e-01
+1.990177854895591736e-02
+1.698033884167671204e-02
+-3.050247021019458771e-02
+1.130666062235832214e-01
+2.015298046171665192e-02
+-1.044731885194778442e-01
+5.582036823034286499e-02
+6.931944750249385834e-03
+-2.916050888597965240e-02
+-4.082124680280685425e-02
+-7.832813076674938202e-03
+9.896343201398849487e-02
+7.808592170476913452e-02
+-7.288336008787155151e-02
+7.714473456144332886e-02
+-1.301961839199066162e-01
+1.848527044057846069e-01
+-9.081019461154937744e-02
+-6.976915174163877964e-04
+2.244408428668975830e-02
+1.066876295953989029e-02
+-1.295847147703170776e-01
+-1.634128987789154053e-01
+-1.808844506740570068e-02
+1.170610636472702026e-01
+-1.211571842432022095e-01
+1.830134838819503784e-01
+-7.467282563447952271e-02
+-6.085983291268348694e-02
+1.003524512052536011e-01
+-5.922634620219469070e-03
+-2.771023809909820557e-01
+2.167193889617919922e-01
+2.497787624597549438e-01
+8.387377485632896423e-03
+-8.958818018436431885e-02
+8.705596625804901123e-02
+4.273151606321334839e-02
+7.838320732116699219e-02
+1.398132294416427612e-01
+9.128690510988235474e-02
+-4.772248119115829468e-02
+1.197547689080238342e-01
+1.196535304188728333e-01
+-4.464437440037727356e-02
+-3.807823732495307922e-02
+7.059801369905471802e-02
+-7.576800137758255005e-02
+2.559782005846500397e-02
+-9.628463536500930786e-02
+-1.367197372019290924e-02
+-6.887884438037872314e-02
+6.069872900843620300e-02
+-8.091127872467041016e-02
+-2.887403964996337891e-02
+-9.738531708717346191e-02
+6.537929177284240723e-02
+-1.219131425023078918e-01
+-5.297732725739479065e-02
+2.407472766935825348e-02
+-2.052211947739124298e-02
+-3.601079434156417847e-02
+1.621966511011123657e-01
+1.322714388370513916e-01
+-3.138993680477142334e-01
+-4.285698011517524719e-02
+-3.461194783449172974e-02
+-2.725695259869098663e-02
+1.069103553891181946e-01
+-9.631482884287834167e-03
+1.439846958965063095e-02
+2.368983924388885498e-01
+-5.414661020040512085e-02
+4.547450318932533264e-02
+-1.666521281003952026e-01
