+++ lib/ClapeyronHANNA/data/HANNA/alpha_2_b.csv
@@ -1,0 +1,96 @@
+-1.749473661184310913e-01
+-6.538887321949005127e-02
+-4.597627744078636169e-02
+-4.625471681356430054e-02
+-4.164475202560424805e-02
+-7.217390090227127075e-02
+6.285990122705698013e-03
+5.774033442139625549e-02
+1.211028844118118286e-01
+-1.746107265353202820e-02
+6.873424351215362549e-02
+1.408745050430297852e-01
+-3.120980225503444672e-02
+4.535458236932754517e-02
+3.821295872330665588e-02
+-1.900108903646469116e-02
+5.711310729384422302e-02
+9.775545448064804077e-02
+-3.209345787763595581e-02
+-1.395642459392547607e-01
+-1.261994987726211548e-01
+-1.880378276109695435e-01
+-1.289127171039581299e-01
+-2.425884455442428589e-02
+6.836225837469100952e-02
+-2.399811297655105591e-01
+-1.930527240037918091e-01
+-7.706368714570999146e-02
+1.708005815744400024e-01
+1.456884890794754028e-01
+-3.568214178085327148e-02
+-7.704883068799972534e-02
+-2.498830668628215790e-02
+1.391231641173362732e-02
+6.622309982776641846e-02
+1.556316111236810684e-02
+1.237079724669456482e-01
+5.212845280766487122e-02
+5.172758176922798157e-02
+-6.275913864374160767e-02
+-6.189687922596931458e-02
+1.023008022457361221e-02
+1.525277793407440186e-01
+-2.088994234800338745e-01
+3.169975057244300842e-02
+-3.149764239788055420e-02
+1.375255081802606583e-02
+-1.653736233711242676e-01
+2.966009452939033508e-02
+5.471393838524818420e-02
+8.921553194522857666e-02
+7.606087625026702881e-02
+1.018550768494606018e-01
+-2.007782906293869019e-01
+-3.127440810203552246e-02
+-1.186676397919654846e-01
+-7.513850927352905273e-02
+1.173744872212409973e-01
+-1.133296564221382141e-01
+1.518554519861936569e-03
+3.047907166182994843e-02
+2.885270118713378906e-02
+7.306510955095291138e-02
+-1.777343451976776123e-02
+-1.431236863136291504e-01
+-1.147060692310333252e-01
+-1.377991586923599243e-01
+1.563341915607452393e-01
+6.867565959692001343e-02
+-8.340056985616683960e-02
+-7.025842368602752686e-02
+4.917898774147033691e-02
+-2.216405794024467468e-02
+-1.003610864281654358e-01
+-4.094214364886283875e-02
+4.080831632018089294e-02
+4.270265996456146240e-02
+8.056865632534027100e-02
+-1.708201467990875244e-01
+1.870822906494140625e-01
+7.113331556320190430e-02
+-1.177071630954742432e-01
+8.170294016599655151e-02
+-6.494610011577606201e-02
+9.439171105623245239e-02
+1.400311142206192017e-01
+8.044847846031188965e-02
+-7.485719770193099976e-02
+4.696375131607055664e-02
+-2.643457241356372833e-02
+1.612265706062316895e-01
+-1.777438819408416748e-01
+1.251357793807983398e-01
+3.423131257295608521e-02
+-3.381965681910514832e-02
+2.364075370132923126e-02
