+++ lib/ClapeyronHANNA/data/HANNA/phi_1_b.csv
@@ -1,0 +1,96 @@
+-1.289850026369094849e-01
+9.679801017045974731e-02
+-5.159948393702507019e-02
+8.957901597023010254e-02
+-1.169027686119079590e-01
+9.904527664184570312e-02
+1.790683250874280930e-03
+-6.504456698894500732e-02
+-2.095818286761641502e-03
+-1.805793195962905884e-01
+-5.095995217561721802e-02
+-2.041716687381267548e-02
+7.844042847864329815e-04
+-5.824472755193710327e-03
+-1.930107921361923218e-01
+5.486955493688583374e-02
+8.685883879661560059e-02
+-1.885851174592971802e-01
+1.030855253338813782e-01
+-1.937486231327056885e-01
+-3.721232712268829346e-02
+-1.487375497817993164e-01
+5.149943381547927856e-02
+-6.424003839492797852e-02
+-1.685919589363038540e-03
+7.419624179601669312e-02
+2.011856809258460999e-02
+-7.108221203088760376e-02
+1.284514963626861572e-01
+-1.244238391518592834e-01
+8.063323050737380981e-02
+1.163796558976173401e-01
+-7.592890411615371704e-02
+6.024687364697456360e-02
+1.173229962587356567e-01
+1.506577730178833008e-01
+-6.617827806621789932e-03
+1.820738762617111206e-01
+1.396837644279003143e-02
+-1.397209912538528442e-01
+1.324573010206222534e-01
+-6.029120087623596191e-02
+8.717847615480422974e-02
+-4.729531332850456238e-02
+6.166252866387367249e-02
+2.365912683308124542e-02
+3.509765490889549255e-02
+-1.265370696783065796e-01
+-6.764986366033554077e-02
+1.272337138652801514e-01
+-6.487507373094558716e-02
+1.586807221174240112e-01
+3.724079579114913940e-02
+-6.967584788799285889e-02
+1.410674210637807846e-02
+-2.226268351078033447e-01
+-3.997076675295829773e-02
+1.014232411980628967e-01
+3.607187792658805847e-02
+-2.185493521392345428e-02
+1.874294131994247437e-01
+-4.594606347382068634e-03
+4.584605619311332703e-02
+8.600047975778579712e-02
+-1.143905520439147949e-01
+5.487243831157684326e-02
+9.428414702415466309e-02
+4.324673116207122803e-02
+-2.133309543132781982e-01
+3.244163095951080322e-02
+2.088101208209991455e-01
+-1.996124684810638428e-01
+1.079291403293609619e-01
+1.273636668920516968e-01
+-4.035928100347518921e-02
+2.805835567414760590e-02
+-1.425755172967910767e-01
+-1.505439132452011108e-01
+1.210136935114860535e-01
+1.466666162014007568e-01
+-1.564974933862686157e-01
+-1.329667270183563232e-01
+-2.222327142953872681e-01
+-5.777566507458686829e-02
+-6.172660365700721741e-02
+-1.002475060522556305e-02
+-3.889885172247886658e-02
+3.082730807363986969e-02
+-3.004297800362110138e-02
+1.449600905179977417e-01
+-1.265668421983718872e-01
+-1.031896993517875671e-01
+2.780663594603538513e-02
+7.270261645317077637e-03
+-3.656169772148132324e-02
+-2.216698527336120605e-01
